backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argcomplete==1.10.3
bcrypt==4.3.0
beautifulsoup4==4.8.2
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
compressed-rtf==1.0.7
cryptography==45.0.4
docx2txt==0.9
ebcdic==1.1.1
extract-msg==0.28.7
Flask==3.1.0
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-genai==1.11.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.72.0rc1
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.22.0
httpx==0.28.1
idna==3.10
IMAPClient==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
PyJWT==2.8.0
lxml==5.4.0
MarkupSafe==3.0.2
nltk==3.9.1
npm==0.1.1
numpy==2.2.4
olefile==0.47
optional-django==0.1.0
packaging==25.0
pandas==2.2.3
pdfminer.six==20191110
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.3
pydantic_core==2.33.1
pygame==2.6.1
PyMuPDF==1.25.5
pyparsing==3.2.3
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-pptx==0.6.23
pytz==2025.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.2
six==1.12.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
SpeechRecognition==3.8.1
SQLAlchemy==2.0.41
textract==1.6.5
threadpoolctl==3.6.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.1.1
urllib3==2.4.0
websockets==15.0.1
Werkzeug==3.1.3
xlrd==1.2.0
XlsxWriter==3.2.3
